social_media_api/requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.1.7
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0
